Add intitialize method to wallet. Also changed how to transaction is being written to block.

+++ src/mipt/infosec/bitcoin/gui/NewJFrame.java
@@ -108,7 +108,7 @@ public class NewJFrame extends javax.swing.JFrame {
         jLabel2.setForeground(new java.awt.Color(255, 0, 0));
 ///////////////        
         Wallet wal = new Wallet();
-        Integer x = wal.getSumm();
+        Double x = wal.getSumm();
         jLabel2.setText(x.toString());
 ////////////////
         jMenu1.setText("Файл");
